requirements.txt
numpy
scikit-learn
openai
tiktoken
anthropic
numba
cohere
chonkie
tenacity
python-dotenv
